Bump to 1.68, fix Clippy warnings, change build
Ultimate goal: bumping toolchain rev to 1.68.

The continuous build was using a separate notion of MSRV (which happened
to equal the pinned toolchain but needed to be updated separately), so
the original bump didn't pass. I have rearranged that build after
talking with @steveklabnik -- the original intent was to build on both
pinned and latest-stable, and it now does that without needing a
separate MSRV variable.

I have, however, added an MRSV to Cargo.toml to give better feedback to
people who try and build Humility with a too-old toolchain.

Once I got past that hurdle, I hit the build that requires Humility to
be clippy clean with the pinned toolchain -- and evidently it wasn't.
The things it reported were absolutely lints in previous versions of
clippy -- I guess they got missed because of some bugs in clippy that
got fixed? Anyway, I have gone through and fixed them all, generally in
the most minimal way possible, but making a related nearby cleanup in a
few cases to simplify the code.

This seems to work now.

Note that the 1.68 toolchain is now warning us about some deps
containing deprecated code, which we should probably fix.

+++ cmd/rpc/src/lib.rs
@@ -337,7 +337,7 @@ impl<'a> RpcClient<'a> {
         let header = RpcHeader {
             image_id: U64::from_bytes(our_image_id.try_into().unwrap()),
             task: U16::new(op.task.task().try_into().unwrap()),
-            op: U16::new(op.code as u16),
+            op: U16::new(op.code),
             nreply: U16::new(nreply as u16),
             nbytes: U16::new(payload.len().try_into().unwrap()),
         };
